+++ lib/src/components/components.dart
@@ -1,5 +1,8 @@
+export 'action_list/octo_action_list.dart';
+export 'action_list/octo_action_list_item.dart';
 export 'button/octo_button.dart';
 export 'flash/octo_flash.dart';
 export 'icon_button/octo_icon_button.dart';
 export 'label/octo_label.dart';
 export 'text_field/octo_text_field.dart';
+export 'tooltip/octo_tooltip.dart';